revert client to use types from audio-filters-web
