+++ packages/audio-filters-web/package.json
@@ -31,12 +31,8 @@
   "devDependencies": {
     "@rollup/plugin-replace": "^6.0.2",
     "@rollup/plugin-typescript": "^12.1.2",
-    "@stream-io/video-client": "workspace:*",
     "rimraf": "^6.0.1",
     "rollup": "^4.40.2",
     "typescript": "^5.8.3"
-  },
-  "peerDependencies": {
-    "@stream-io/video-client": "*"
   }
 }
