+++ packages/client/src/devices/MicrophoneManager.ts
@@ -1,4 +1,5 @@
 import { combineLatest, Observable } from 'rxjs';
+import type { INoiseCancellation } from '@stream-io/audio-filters-web';
 import { Call } from '../Call';
 import { InputMediaDeviceManager } from './InputMediaDeviceManager';
 import { MicrophoneManagerState } from './MicrophoneManagerState';
@@ -19,7 +20,6 @@ import {
 } from '../store/rxUtils';
 import { RNSpeechDetector } from '../helpers/RNSpeechDetector';
 import { withoutConcurrency } from '../helpers/concurrency';
-import type { INoiseCancellation } from '../types';
 
 export class MicrophoneManager extends InputMediaDeviceManager<MicrophoneManagerState> {
   private speakingWhileMutedNotificationEnabled = true;
@@ -78,17 +78,7 @@ export class MicrophoneManager extends InputMediaDeviceManager<MicrophoneManager
 
         if (autoOn && callingState === CallingState.JOINED) {
           this.noiseCancellationRegistration
-            .then(() => {
-              if (this.noiseCancellation?.canAutoEnable) {
-                return this.noiseCancellation.canAutoEnable();
-              }
-              return true;
-            })
-            .then((canAutoEnable) => {
-              if (canAutoEnable) {
-                this.noiseCancellation?.enable();
-              }
-            })
+            .then(() => this.noiseCancellation?.enable())
             .catch((err) => {
               this.logger('warn', `Failed to enable noise cancellation`, err);
               return this.call.notifyNoiseCancellationStopped();
@@ -107,9 +97,14 @@ export class MicrophoneManager extends InputMediaDeviceManager<MicrophoneManager
   /**
    * Enables noise cancellation for the microphone.
    *
+   * Note: not supported in React Native.
    * @param noiseCancellation - a noise cancellation instance to use.
    */
   async enableNoiseCancellation(noiseCancellation: INoiseCancellation) {
+    if (isReactNative()) {
+      throw new Error('Noise cancellation is not supported in React Native');
+    }
+
     const { ownCapabilities, settings } = this.call.state;
     const hasNoiseCancellationCapability = ownCapabilities.includes(
       OwnCapability.ENABLE_NOISE_CANCELLATION,
@@ -144,17 +139,12 @@ export class MicrophoneManager extends InputMediaDeviceManager<MicrophoneManager
         },
       );
 
-      if (isReactNative()) {
-        // no filter registration in React Native, its done in the native code on app init
-        this.noiseCancellationRegistration = Promise.resolve();
-      } else {
-        const registrationResult = this.registerFilter(
-          noiseCancellation.toFilter(),
-        );
-        this.noiseCancellationRegistration = registrationResult.registered;
-        this.unregisterNoiseCancellation = registrationResult.unregister;
-        await this.noiseCancellationRegistration;
-      }
+      const registrationResult = this.registerFilter(
+        noiseCancellation.toFilter(),
+      );
+      this.noiseCancellationRegistration = registrationResult.registered;
+      this.unregisterNoiseCancellation = registrationResult.unregister;
+      await this.noiseCancellationRegistration;
 
       // handles an edge case where a noise cancellation is enabled after
       // the participant as joined the call -> we immediately enable NC
@@ -163,13 +153,7 @@ export class MicrophoneManager extends InputMediaDeviceManager<MicrophoneManager
           NoiseCancellationSettingsModeEnum.AUTO_ON &&
         this.call.state.callingState === CallingState.JOINED
       ) {
-        let canAutoEnable = true;
-        if (noiseCancellation.canAutoEnable) {
-          canAutoEnable = await noiseCancellation.canAutoEnable();
-        }
-        if (canAutoEnable) {
-          noiseCancellation.enable();
-        }
+        noiseCancellation.enable();
       }
     } catch (e) {
       this.logger('warn', 'Failed to enable noise cancellation', e);
@@ -181,8 +165,13 @@ export class MicrophoneManager extends InputMediaDeviceManager<MicrophoneManager
 
   /**
    * Disables noise cancellation for the microphone.
+   *
+   * Note: not supported in React Native.
    */
   async disableNoiseCancellation() {
+    if (isReactNative()) {
+      throw new Error('Noise cancellation is not supported in React Native');
+    }
     await (this.unregisterNoiseCancellation?.() ?? Promise.resolve())
       .then(() => this.noiseCancellation?.disable())
       .then(() => this.noiseCancellationChangeUnsubscribe?.())
