+++ packages/noise-cancellation-react-native/package.json
@@ -48,7 +48,6 @@
   "homepage": "https://github.com/GetStream/stream-video-js#readme",
   "devDependencies": {
     "@stream-io/react-native-webrtc": "^125.3.0",
-    "@stream-io/video-client": "workspace:*",
     "react-native": "^0.79.2",
     "react-native-builder-bob": "^0.37.0",
     "rimraf": "^6.0.1",
@@ -56,7 +55,6 @@
   },
   "peerDependencies": {
     "@stream-io/react-native-webrtc": ">=125.3.0",
-    "@stream-io/video-client": "*",
     "react-native": "*"
   },
   "react-native-builder-bob": {
